requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.23
aiosqlite==0.19.0
aiofiles==23.2.1
matplotlib==3.8.2
python-dotenv==1.0.0